+++ Davidian/Engine/CMakeLists.txt
@@ -11,6 +11,11 @@ set(SOURCES
         OrbitState
         )
 
+set(TEST_SOURCES
+        CelestialSystemTest
+        CSFileManagerTest
+        OrbitalEngineTest OrbitalHistoryTest.cpp OrbitStateTest.cpp)
+
 if(BUILD_TESTING)
     list(APPEND SOURCES UnitTestHelpers)
 endif()
@@ -25,6 +30,6 @@ target_include_directories(${PROJECT_NAME} PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
 
 if (BUILD_TESTING)
     target_include_directories(${PROJECT_NAME} PRIVATE ${GTEST_INCLUDE_DIR} ${GMOCK_INCLUDE_DIR})
-    add_executable(${PROJECT_NAME}_TEST ${SOURCES})
+    add_executable(${PROJECT_NAME}_TEST ${TEST_SOURCES})
     target_link_libraries(${PROJECT_NAME}_TEST gmock gmock_main gtest ${PROJECT_NAME})
 endif()
